修正

+++ texture_ver/srcs/5_raycaster_draw.c
@@ -6,7 +6,7 @@
 /*   By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 23:41:22 by yokitaga          #+#    #+#             */
-/*   Updated: 2023/11/20 00:38:16 by yokitaga         ###   ########.fr       */
+/*   Updated: 2023/11/20 00:45:37 by yokitaga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ int	key_draw(t_data *data)
 	data->img->addr = mlx_get_data_addr(data->img->img, &data->img->bits_per_pixel, &data->img->size_line, &data->img->endian);
 	draw_celling_fllor(data);
 	row = 0;
-	while (row < (unsigned int)SCREEN_W)
+	while (row < SCREEN_W)
 	{
 		data->camera_x = 2 * row / (double)SCREEN_W - 1;
 		calc_init(data);
